refactor(model): restructure application_command module
This PR flattens `model::application::interaction::application_command`,
which makes it easier to navigate through the modules.
