+++ model/src/application/interaction/application_command/resolved.rs
